Move integration tests to own files
